split message handlers from SoulseekClient
